Consolidate config and provisioner files
Prior to this commit, there were directories for config and provisioner,
but only one file in each of them. This commit consolidates those files
to simplify the plugin layout.
